+++ mainwindow.h
@@ -99,7 +99,7 @@ public slots:
     void updateMemories();
     void mainTabChanged(int value);
     void clearTextArea();
-    void updateFreqGUI(long long center_freq, long carrier_offset);
+    void updateFreqGUI(long long center_freq, long long carrier_offset);
     void enterShift();
     void updateRxCTCSS(int value);
     void updateTxCTCSS(int value);
@@ -171,7 +171,7 @@ signals:
     void tuneFreq(qint64 center_freq);
     void tuneTxFreq(qint64 freq);
     void changeTxShift(qint64 center_freq);
-    void fineTuneFreq(long center_freq);
+    void fineTuneFreq(long long center_freq);
     void toggleWidebandMode(bool value);
     void toggleRxModemMode(int value);
     void toggleTxModemMode(int value);
