+++ src/commandprocessor.h
@@ -75,7 +75,7 @@ signals:
     void tuneFreq(qint64 center_freq);
     void tuneTxFreq(qint64 freq);
     void changeTxShift(qint64 center_freq);
-    void fineTuneFreq(long center_freq);
+    void fineTuneFreq(long long center_freq);
 
     /// Transmit
     void startTransmission();
